% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitividad.R
\name{competitive}
\alias{competitive}
\title{Competitividad // Competitiveness}
\usage{
competitive(data, nivel = "provincia")
}
\arguments{
\item{data}{un tibble guardado como objeto en el Enviroment luego de consultar `election_get()` // 
A tibble saved as an object in the Enviroment after querying `election_get ()`}

\item{nivel}{establece el nivel de desagregacion sobre el que se quiere calcular la competitividad: por defualt es 'provincia' y se desagregan las observaciones asignando los valores 'departmento'  o 'circuito' al parametro. // 
Wstablishes the level of disaggregation on which you want to calculate competitiveness: by definition it is 'province' and the observations are disaggregated by assigning the values 'department' or 'circuit' to the parameter.}
}
\description{
Funcion que devuelve un valor entre 0 y 1 que describe el grado de competencia de una eleccion en un distrito determinado // 
Function that returns a value between 0 and 1 that describes the degree of competition of an election in a given district
}
