% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiveness.R
\name{compute_competitiveness}
\alias{compute_competitiveness}
\title{Competitividad // Competitiveness}
\usage{
compute_competitiveness(data, level = "provincia")
}
\arguments{
\item{data}{un tibble guardado como objeto en el Enviroment luego de consultar `election_get()` // 
A tibble saved as an object in the Enviroment after querying `election_get ()`}

\item{level}{establece el level de desagregacion sobre el que se quiere calcular la competitividad: por defualt es 'provincia' 
y se desagregan las observaciones asignando los valores 'departmento'  o 'circuito' al parametro. // 
Establishes the level of aggregation on which you want to compute competitiveness: by definition it is 'province' 
and the observations are disaggregated by assigning the values 'departamento' or 'circuito' to the parameter.}
}
\description{
Funcion que devuelve un valor entre 0 y 1 que describe el grado de competencia de una eleccion en un distrito determinado // 
Function that returns a value between 0 and 1 that describes the degree of competition of an election in a given district
}
