% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_disproportion.R
\name{compute_disproportion}
\alias{compute_disproportion}
\title{Calcula Desproporción del Sistema de Partidos 
 (\emph{Computes Parties System Disproportion})}
\usage{
compute_disproportion(data, formula = "gallagher")
}
\arguments{
\item{data}{la base de datos para hacer el cálculo obtenida con \code{\link{get_election_data}} 
(\emph{tiblle downloaded with \code{\link{get_election_data}} needed to compute disproportion}).}

\item{formula}{variante de cálculo elegido ('gallagher', 'cox_shugart', 'lijphart' o 'rae'). Por defecto es la formula de Gallagher}
}
\value{
Devuelve un tibble con \code{class "tbl_df","tbl", "data.frame"} con el cómputo
 (\emph{Returns a tibble with \code{class "tbl_df", "tbl", "data.frame"} with the computation}).
}
\description{
Función que calcula el índide de desprorpoción del sistema de partidos midiendo la distancia relativa entre proporción de votos
y de bancas obtendios por los partidos políticos (Gallagher por defecto)
 (\emph{Function that computes the party system disproportion index})
}
\details{
\deqn{Rae = \frac{\sum |V_{i} - S_{i} |}{n}}

\deqn{Lijphart = max( |V_{i} - S_{i}| )}

\deqn{Gallagher =\sqrt{1/2 \sum_{i}^{n}(V_{i} - S_{i})^2}}

\deqn{Cox - Shugart :\tfrac{V_{i} \alpha}{S_{i}} = \beta}

\strong{REQUISITO:}

El formato de \code{data} debe ser \code{long} para calcular \code{\link{compute_disproportion}}. 
 Si \code{data} es \emph{wide} se puede transformar con \code{\link{make_long}}  
 (\emph{\code{long} format of \code{data} is required for \code{\link{compute_disproportion}}. 
 If \code{data} is in \emph{wide} format you can transform it with \code{\link{make_long}}})

\strong{NOTA:}

para calcular el grado de desprorpoción el parámetro \code{level} de \code{\link{get_election_data}} debe ser "provincia" (\emph{default})  
 (\emph{To compute the degree of disproportion, the parameter \code{level} of \code{\link{get_election_data}} must be "provincia" (\emph{default}}).
}
\examples{

tucuman_dip_gral_2017

tucuman_dip_gral_2017 \%>\% 
  compute_disproportion()

}
\seealso{
\code{\link{compute_competitiveness}, \link{compute_seats},\link{compute_nep}}
}
