% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionario_elecciones.R
\name{diccionario_elecciones}
\alias{diccionario_elecciones}
\title{Diccionario de Elecciones}
\usage{
diccionario_elecciones(viewer = FALSE)
}
\arguments{
\item{viewer}{Por default es `FALSE`. Cuando `TRUE` devuelve una tabla en el Viewer}
}
\description{
Funcion que devuelve un tibble con los parametros para llamar elecciones con funcion `election_get()`
}
