% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_get.R
\name{election_get}
\alias{election_get}
\title{Descarga de Bases de resultados electorales - nivel mesa}
\usage{
election_get(distrito, categoria, turno, anio)
}
\arguments{
\item{distrito}{un character con mobre con codigo para argentina y las 24 provincias}

\item{turno}{un integer para el anio de eleccion}

\item{cargo}{un character para la categoria electoral - diputado, senador o presidente: dip , sen, presi}

\item{tipo}{tipo de eleccion - primaria o general: paso, gral}
}
\description{
Funcion que descarga bases de resultados electorales nacionales desde el año 2007
}
