% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_votes.R
\name{get_bill_votes}
\alias{get_bill_votes}
\title{Obtener votos de un proyecto de ley 
(\emph{Download bill vote data})}
\usage{
get_bill_votes(bill = NULL)
}
\arguments{
\item{bill}{Parametro en el que se especifica el id del proyecto obtenido con \code{\link{show_available_bills}}}
}
\description{
Función que devuelve un \emph{data.frame} con el detalle de los votos individuales de los legisladores a un proyecto de ley.
}
\examples{

get_bill_votes(bill = "1926-Diputados")

}
\seealso{
\code{\link{plot_bill}, \link{show_available_bills}}
}
