% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_listas.R
\name{get_lista}
\alias{get_lista}
\title{Obtener nombres de listas}
\usage{
get_lista(data)
}
\arguments{
\item{data}{un tibble guardado como objeto en el Enviroment luego de consultar `election_get()` habiendo seleccionado parametro `long = TRUE`}
}
\description{
Funcion que agrega el nombre de las listas a un tibble obtenido con `election_get()`
}
