% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long.R
\name{get_long}
\alias{get_long}
\title{Transforma tibble a formato largo // transform tibble into long format}
\usage{
get_long(data)
}
\arguments{
\item{data}{es el tibble que devuelve `election_get(data = data, LONG = FALSE)`// `election_get (data = data, LONG = FALSE)` output tibble}
}
\description{
Funcion auxiliar que transforma el tibble WIDE obtenido con `election_get()` a LONG // Auxiliary function that transforms a WIDE tibble obtained with `election_get ()` to LONG format
}
