% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Obtiene nombres de listas (\emph{Get party names})}
\usage{
get_names(data)
}
\arguments{
\item{data}{un tibble descargado con \code{\link{get_election_data}(long = TRUE)} guardado como objeto en el Enviroment
(\emph{A \code{\link{get_election_data}(long = TRUE)} tibble saved as an object in the Enviroment}).}
}
\value{
Devuelve el data set original con una columna extra con la identifiacion de las listas o partidos politicos. 
Como el objeto de entrada, este es \code{class "tbl_df","tbl","data.frame"} 
(\emph{it retruns the original data set with a binded column with political parties names. 
 As the original input the object is of \code{class "tbl_df","tbl","data.frame"}}).
}
\description{
Función que agrega el nombre de las listas o partidos como columna a un tibble obtenido con \code{\link{get_election_data}(long = TRUE)} 
(\emph{Function that adds party labels as a column to a tibble obtained with \code{\link{get_election_data}(long = TRUE)}}).
}
\details{
El formato de \code{data} debe ser \code{long} para poder obtener 
nombres de listas con \code{\link{get_names}}. Si \code{data} es \emph{wide} se puede transformar con \code{\link{make_long}} 
(\emph{\code{long} format of \code{data} is required to get party labels with \code{\link{get_names}}.  If \code{data} is in
\emph{wide} format you can transform it with \code{\link{make_long}}}).
}
\examples{
 
 tucuman_dip_gral_2017
 
 tucuman_dip_gral_2017 \%>\% 
     get_names()
   
}
