% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_results.R
\name{map_results}
\alias{map_results}
\title{Mapea resultados (\emph{Map results})}
\usage{
map_results(data)
}
\arguments{
\item{data}{un tibble guardado como objeto en el Enviroment luego de consultar \code{\link{get_election_data}} con parámetro
\code{level} en \code{provincia} para elecciones presidenciales y \code{level} en \code{departamento} para legislativas 
(\emph{tibble saved as an object in the Enviroment after querying \code{\link{get_election_data}}} with \code{provincia} as 
\code{level} parameter for presidential elections and \code{departmanto} for legislative elections). 
#'}
}
\value{
Devuelve un objeto de \code{class"gg" "ggplot"} que mapea el resultado de una eleccion condicional al nivel de agregacion de data 
 (\emph{Returns an object of \code{class "gg" "ggplot"} that plots the election results in a map conditional on the level of \code{data} aggregation.}).
}
\description{
Función para mapear resultados de la elección 
 (\emph{Function to map election results})
}
\details{
\strong{REQUISITOS:}

\strong{1}. El formato de \code{data} debe ser \code{long} para poder graficar. Si \code{data} es \emph{wide} se puede 
 transformar con \code{\link{make_long}} 
 (\emph{\code{long} format of \code{data} is required for plotting results. If \code{data} is in \emph{wide} format you can transform 
 it with \code{\link{make_long}}})

\strong{2.} \code{data} tiene que haber incorporando los nombres de las listas. Agreguelos con \code{\link{get_names}} 
 (\emph{\code{data} must have party names. Add them with \code{\link{get_names}}})
}
\seealso{
\code{\link{tabulate_results}, \link{plot_results}}
}
