% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nep.R
\name{nep}
\alias{nep}
\title{Calcla Número Efectivo de Partidos - NEP}
\usage{
nep(data, index = "All")
}
\arguments{
\item{data}{la base de datos para hacer el calculo descargada con 'election_get' - relevante el 'nivel' de agregación sobre la que se hará el calculo: provincia, departamento o circuito}

\item{index}{un character con la fórmula elegida: "Laakso-Taagepera" o "Golosov"}
}
\description{
Funcion que calcula el Numero Efectivo de Partidos - NEP. Indicador que provee un número "ajustado" de partidos políticos en un sistema de partídos
}
