% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_speech.R
\name{plot_speech}
\alias{plot_speech}
\title{Visualización rápida del discurso
 (\emph{Quick speech viz})}
\usage{
plot_speech(data)
}
\arguments{
\item{data}{requiere un discurso dentro data.frame con formato \emph{tidy} descargado con \code{get_speech} 
\emph{(a tidy data.frame with the speech downloaded with \code{get_speech} is required}).}
}
\value{
Devuelve un objeto con clases \code{"wordcloud2" "htmlwidget"} que representa gráficamente el contenido de un discurso presidencial  
 (\emph{it retruns an object of \code{classes "wordcloud2" "htmlwidget"}} with a graphical representation of presidential speeches).
}
\description{
Función que permite una rápida visualización de los discursos presidenciales a través de nube de palabras representando su frecuencia relativa con el tamaño. 
 (\emph{Function that plots a word cloud of presidential specches, with word sizes as a function of its frequency in the data set})
}
