% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\name{polar}
\alias{polar}
\title{Draw polar plot}
\usage{
polar(data1, data2, dot_alpha = 1, seg_size = 1, seg_alpha = dot_alpha)
}
\arguments{
\item{data1}{a list of dots in a `data.frame` with following columns:
- x
- dot_shape
- dot_color
- dot_size}

\item{data2}{a list of connection in a `data.frame` with following columns:
- x1
- x2
- segment_color}

\item{dot_alpha}{alpha for dot}

\item{seg_size}{size for segment}

\item{seg_alpha}{alpha for segment}
}
\value{
a `ggplot`
}
\description{
Draw polar plot
}
\examples{
data1 <- data.frame(
  x = letters[1:7],
  dot_shape = c("r", "r", "r", "b", "b", "b", "b"),
  dot_color = c("r", "r", "r", "b", "b", "b", "b"),
  dot_size = abs(rnorm(7))
)

data2 <- data.frame(
  x1 = letters[1:7],
  x2 = c("b", "c", "d", "e", "c", "a", "c"),
  seg_color = c("r", "r", "r", "b", "b", "b", "b")
)

polar(data1, data2)

}
