% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpolar.R
\name{polar_connect}
\alias{polar_connect}
\title{Connects dots}
\usage{
polar_connect(data, x1, x2, ...)
}
\arguments{
\item{data}{a \code{data.frame} contains connections of all events.}

\item{x1, x2}{the column names (without quote) storing connected events.}

\item{...}{parameters passing to \link[ggplot2:geom_segment]{ggplot2::geom_segment},
expect \code{c(x, xend, y, yend)} these 4 mapping parameters.}
}
\value{
a \code{ggplot} object.
}
\description{
Check \code{\link[=polar_init]{polar_init()}} for examples.
}
