% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secciones_pba.R
\docType{data}
\name{secciones_pba}
\alias{secciones_pba}
\title{Base de datos con id de secciones electorales de la Provincia de Buenos Aires por departamento}
\format{
Un data frame con 135 filas y 2 variables:
\describe{
  \item{coddepto}{Código de Departamento - 135}
  \item{seccion}{Código de Sección Electoral - de utilidad para geofacet plot}
  ...
}
}
\usage{
secciones_pba
}
\description{
Base de datos con id de secciones electorales de la Provincia de Buenos Aires por departamento
}
\keyword{datasets}
