% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_bills.R
\name{show_available_bills}
\alias{show_available_bills}
\title{Diccionario de proyectos de ley sometidos a votación 
(\emph{Bills collection})}
\usage{
show_available_bills(viewer = TRUE)
}
\arguments{
\item{viewer}{Por default es \code{TRUE} y  muestra una tabla formateada en el \emph{Viewer} de \emph{RStudio}. Cuando \code{FALSE} imprime en consola.}
}
\description{
Función que devuelve un \emph{data.frame} listado de proyectos de ley con un id por proyecto y cámara legislativa
}
\examples{

show_available_bills(viewer = FALSE)


}
\seealso{
\code{\link{plot_bill}, \link{get_bill_votes}}
}
