% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_elections.R
\name{show_available_elections}
\alias{show_available_elections}
\title{Diccionario de Elecciones disponibles 
 (\emph{Elections Collection})}
\usage{
show_available_elections(viewer = FALSE)
}
\arguments{
\item{viewer}{Por default es \code{FALSE}. Cuando \code{TRUE} devuelve una tabla en el Viewer de RStudio 
(\emph{The default is \code{FALSE}. When \code{TRUE} it returns a table in RStudio Viewer}).}
}
\description{
Funcion que devuelve un tibble con los parametros necesarios para llamar elecciones con \code{\link{get_election_data}} 
 (\emph{Function that returns a tibble with the necessary parameters to call elections with} \code{\link{get_election_data}})
}
