% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_speech.R
\name{show_available_speech}
\alias{show_available_speech}
\title{Diccionario de discursos disponibles
 (\emph{Speeches collection})}
\usage{
show_available_speech(viewer = FALSE)
}
\arguments{
\item{viewer}{Por defecto es \code{FALSE}. Cuando \code{TRUE} devuelve una tabla en el \emph{Viewer} de \emph{RStudio} 
(\emph{The default is \code{FALSE}. When \code{TRUE} it returns a table in \emph{RStudio Viewer}}).}
}
\value{
El objeto de salida es un data set con los id dediscursos disponibles para usar como parámetro con
 con \code{\link{get_speech}}. Cuando el parámetro es \code{viewer = FALSE}, devuelve un tibble con \code{class "tbl_df","tbl","data.frame"}, y 
 cuando es \code{viewer = TRUE} devuelve un objeto con \code{class "datatables","htmlwidget"} 
 (\emph{The output is a data set with speeches id needed as parameters in \code{\link{get_speech}}. 
 When parameter is set to \code{viewer = FALSE} it returns a tibble and when it is \code{viewer = TRUE} it returns an 
 object of \code{class "datatables","htmlwidget"}}).
}
\description{
Función que devuelve un listado de discursos de apertura de sesiones emitidos por los presidentes de Argentina ante la Asamblea Legislativa.
}
\examples{
 
 show_available_speech()
 

}
