% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tucuman_dip_gral_2017.R
\docType{data}
\name{tucuman_dip_gral_2017}
\alias{tucuman_dip_gral_2017}
\title{Data set de muestra}
\format{Un data frame con 6 filas y 8 variables:
\describe{
  \item{category}{Categoria Electoral}
  \item{round}{Ronda o Truno Electoral}
  \item{year}{Fecha de eleccion}
  \item{codprov}{Codigo de Provincia}
  \item{name_prov}{Nombre de Provincia}
  \item{electores}{Cantidad de Electores}
  \item{listas}{Codigo de lista o partido}
  \item{votos}{Cantidad de votos absolutos}
  ...
}}
\usage{
tucuman_dip_gral_2017
}
\description{
Un tibble con los resultados agregados de la eleccion general para Diputado Nacional por la provincia de Tucuman en 2017
}
\keyword{datasets}
