% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitividad.R
\name{competitividad}
\alias{competitividad}
\title{Competitividad}
\usage{
competitividad(data, nivel = "distrito")
}
\arguments{
\item{data}{un tibble guardado como objeto en el Enviroment luego de consultar `election_get()`}

\item{nivel}{establece el nivel de desagregaci칩n sobre el que se quiere calcular la competitividad: 'provincial' mostrar치 detalle al interior del distrito. 'departmantal'  lo har치 al nicel de circuitos electorales.}
}
\description{
Funcion que devuelve un valor entre 0 y 1 que describe el grado de competencia de una elecci칩n en un distrito determinado
}
