% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_competitiveness.R
\name{compute_competitiveness}
\alias{compute_competitiveness}
\title{Competitividad  (\emph{Competitiveness})}
\usage{
compute_competitiveness(data, level = "provincia")
}
\arguments{
\item{data}{un tibble guardado como objeto en el Enviroment luego de consultar \code{\link{get_election_data}} 
(\emph{A tibble saved as an object in the Enviroment after querying \code{\link{get_election_data}}}).}

\item{level}{un character que establece el nivel de desagregacion sobre el que se quiere calcular la competitividad: 
por defualt es \code{provincia} y se desagregan las observaciones asignando los valores \code{departamento} o \code{circuito} al parametro
(\emph{Establishes the level of aggregation on which you want to compute competitiveness: by definition it is \code{provincia} 
and the observations are disaggregated by assigning the values \code{departamento} or \code{circuito} to the parameter}).}
}
\description{
Funcion que devuelve un valor entre 0 y 1 que describe el grado de competencia de una eleccion en un distrito determinado  
(\emph{Function that returns a value between 0 and 1 that describes the degree of competition of an election in a given district})
}
\details{
\eqn{Competitividad} mide la diferencia porcentual de votos v치lidos obtenidos por los dos partidos m치s votados \eqn{a} y \eqn{b}. 
Puede tomar valores entre \eqn{[0,1]} donde \eqn{1} es lo m치s comeptitivo (\eqn{a = 50\%} y \eqn{b = 50\%} de los votos).
 La f칩rmula utilizada es:  \deqn{Competitividad = 1 - (a - b)}

\strong{REQUISITO:} El formato de \code{data} debe ser \code{long} para calcular \code{\link{compute_competitiveness}}. 
Si \code{data} es \emph{wide} se puede transformar con \code{\link{get_long}} 
(\emph{\code{long} format of \code{data} is required for \code{\link{compute_competitiveness}}. If \code{data} is in \emph{wide} format
you can transform it with \code{\link{get_long}}}).
}
