% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long.R
\name{get_long}
\alias{get_long}
\title{Transforma tabla de resultados a formato largo}
\usage{
get_long(data)
}
\arguments{
\item{data}{es el tibble que devuelve `election_get(data = data, LONG = FALSE)`}
}
\description{
Funcion auqiliar que transforma el tibble WIDE de `election_get()` a LONG
}
