% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Obtiene nombres de listas (\emph{Get party names})}
\usage{
get_names(data)
}
\arguments{
\item{data}{un tibble descargado con \code{\link{get_election_data} = TRUE} guardado como objeto en el Enviroment
(\emph{An \code{\link{get_election_data} = TRUE} tibble saved as an object in the Enviroment})}
}
\description{
Funcion que agrega el nombre de las listas o partidos como columna a un tibble obtenido con \code{\link{get_election_data} = TRUE} 
(\emph{Function that adds the name of the lists or parties as a column to a tibble obtained with \code{\link{get_election_data} = TRUE}})
}
