% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_listas.R
\name{get_names}
\alias{get_names}
\title{Obtiene nombres de listas // Get party names}
\usage{
get_names(data)
}
\arguments{
\item{data}{un tibble descargado con `election_get (data, long = TRUE)` guardado como objeto en el Enviroment // 
An `election_get(data, long = TRUE)` tibble saved as an object in the Enviroment}
}
\description{
Funcion que agrega el nombre de las listas o partidos como columna a un tibble obtenido con `election_get(data,long = TRUE)` //
Function that adds the name of the lists or parties as a column to a tibble obtained with `election_get(data, long = TRUE)`
}
