% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_global_variables.R
\docType{package}
\name{polAr}
\alias{polAr}
\alias{polAr-package}
\title{\code{polAr} package}
\description{
Caja de Herramientas para el procesamiento de datos electorales de Argentina
See the README on
\href{https://github.com/electorArg/polAr/blob/master/README.md}{Github}
}
