% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections_collection.R
\name{show_available_elections}
\alias{show_available_elections}
\title{Diccionario de Elecciones disponibles // Elections Collection}
\usage{
show_available_elections(viewer = FALSE)
}
\arguments{
\item{viewer}{Por default es `FALSE`. Cuando `TRUE` devuelve una tabla en el Viewer de RStudio // 
The default is `FALSE`. When `TRUE` returns a table in RStudio Viewer}
}
\description{
Funcion que devuelve un tibble con los parametros necesarios para llamar elecciones con`election_get()` // 
Function that returns a tibble with the necessary parameters to call elections with `election_get ()`
}
